Crate spinlock

no_std spin lock implementation that can disable kernel local IRQs or preemption while locking.

Cargo Features

smp: Use in the multi-core environment. For single-core environment (without this feature), the lock state is unnecessary and optimized out. CPU can always get the lock if we follow the proper guard in use. By default, this feature is disabled.

Structs

BaseSpinLock

A spin lock providing mutually exclusive access to data.

BaseSpinLockGuard

A guard that provides mutable data access.

Type Aliases

SpinNoIrq

A spin lock that disables kernel preemption and local IRQs while trying to lock, and re-enables it after unlocking.

SpinNoIrqGuard

A guard that provides mutable data access for SpinNoIrq.

SpinNoPreempt

A spin lock that disables kernel preemption while trying to lock, and re-enables it after unlocking.

SpinNoPreemptGuard

A guard that provides mutable data access for SpinNoPreempt.

SpinRaw

A raw spin lock that does nothing while trying to lock.

SpinRawGuard

A guard that provides mutable data access for SpinRaw.