Struct spinlock::PoisonErrorStable

pub struct PoisonError<T> {
    // some fields omitted
}

A type of error which can be returned whenever a lock is acquired.

Both Mutexes and RwLocks are poisoned whenever a task fails while the lock is held. The precise semantics for when a lock is poisoned is documented on each lock, but once a lock is poisoned then all future acquisitions will return this error.

Methods

impl<T> PoisonError<T>

fn new(guard: T) -> PoisonError<T>

Create a PoisonError.

fn into_guard(self) -> T

Consumes this error indicating that a lock is poisoned, returning the underlying guard to allow access regardless.

fn into_inner(self) -> T

Consumes this error indicating that a lock is poisoned, returning the underlying guard to allow access regardless.

fn get_ref(&self) -> &T

Reaches into this error indicating that a lock is poisoned, returning a reference to the underlying guard to allow access regardless.

fn get_mut(&mut self) -> &mut T

Reaches into this error indicating that a lock is poisoned, returning a mutable reference to the underlying guard to allow access regardless.